restricting deletion of jv when it is utilized

+++ api/src/main/kotlin/com/cogoport/ares/api/settlement/service/implementation/ParentJVServiceImpl.kt
@@ -30,6 +30,7 @@ import com.cogoport.ares.api.settlement.service.interfaces.ParentJVService
 import com.cogoport.ares.api.settlement.service.interfaces.ThirdPartyApiAuditService
 import com.cogoport.ares.api.utils.Util
 import com.cogoport.ares.api.utils.Utilities
+import com.cogoport.ares.api.utils.logger
 import com.cogoport.ares.model.common.ResponseList
 import com.cogoport.ares.model.payment.AccMode
 import com.cogoport.ares.model.payment.AccountType
@@ -51,6 +52,7 @@ import com.cogoport.plutus.model.invoice.SageOrganizationRequest
 import com.cogoport.plutus.model.invoice.SageOrganizationResponse
 import com.fasterxml.jackson.databind.ObjectMapper
 import io.micronaut.context.annotation.Value
+import io.sentry.Sentry
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
 import org.json.JSONObject
@@ -303,26 +305,38 @@ open class ParentJVServiceImpl : ParentJVService {
 
     override suspend fun deleteJournalVoucherById(id: String, performedBy: UUID): String {
         val parentJvId = Hashids.decode(id)[0]
-        parentJVRepository.deleteJournalVoucherById(parentJvId, performedBy)
-        val jvLineItemData = journalVoucherRepository.getJournalVoucherByParentJVId(parentJvId)
-        jvLineItemData.forEach { lineItem ->
-            if (lineItem.status == JVStatus.APPROVED) {
-                accountUtilizationRepository.deleteAccountUtilizationByDocumentValueAndAccType(lineItem.jvNum, AccountType.valueOf(lineItem.category))
+        try {
+            val parentJvDetails = parentJVRepository.findById(parentJvId) ?: throw AresException(AresError.ERR_1002, "JV")
+            if (parentJvDetails.isUtilized == true) {
+                throw AresException(AresError.ERR_1540, "JV is already utilized.")
             }
-        }
-        journalVoucherRepository.deleteJvLineItemByParentJvId(parentJvId, performedBy)
+            parentJVRepository.deleteJournalVoucherById(parentJvId, performedBy)
+            val jvLineItemData = journalVoucherRepository.getJournalVoucherByParentJVId(parentJvId)
+            jvLineItemData.forEach { lineItem ->
+                if (lineItem.status == JVStatus.APPROVED) {
+                    accountUtilizationRepository.deleteAccountUtilizationByDocumentValueAndAccType(lineItem.jvNum, AccountType.valueOf(lineItem.category))
+                }
+            }
+            journalVoucherRepository.deleteJvLineItemByParentJvId(parentJvId, performedBy)
 
-        auditService.createAudit(
-            AuditRequest(
-                objectType = AresConstants.JOURNAL_VOUCHERS,
-                objectId = parentJvId,
-                actionName = AresConstants.DELETE,
-                data = mapOf("id" to id, "status" to "DELETED"),
-                performedBy = performedBy.toString(),
-                performedByUserType = null
+            auditService.createAudit(
+                    AuditRequest(
+                            objectType = AresConstants.JOURNAL_VOUCHERS,
+                            objectId = parentJvId,
+                            actionName = AresConstants.DELETE,
+                            data = mapOf("id" to id, "status" to "DELETED"),
+                            performedBy = performedBy.toString(),
+                            performedByUserType = null
+                    )
             )
-        )
-
+        } catch (aresException: AresException) {
+            logger().error("""${mapOf("data" to id, "error" to "${aresException.context} ${aresException.error.message}")}""")
+            throw aresException
+        } catch (ex: Exception) {
+            logger().error(ex.stackTraceToString())
+            Sentry.captureException(ex)
+            throw ex
+        }
         return id
     }
